RavenDB-20642 : move 'GetPerformanceStats' to etl tests infrastructure
